Создает файл message-form и в него переносит действия с модульным окном сообщения
